app/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
google-cloud-aiplatform>=1.38.1
python-dotenv>=1.0.0
gradio>=4.44.1
google-generativeai